+++ src/gui/Layout.h
@@ -44,7 +44,7 @@ public:
 	/**
 	 * Increases the adjustments by the given amounts
 	 */
-	void scrollRelativ(int x, int y);
+	void scrollRelativ(double x, double y);
 
 	/**
 	 * Changes the adjustments by absolute amounts (for pinch-to-zoom)
